README.md
Clicker Game
This is a simple clicker game implemented in Java using Swing for GUI.

Description
The Clicker Game is a basic application where the user clicks a button to increase a counter displayed on the screen. It serves as a simple example of event-driven programming in Java using Swing components.

Features
Click button to increase a counter
Counter displayed on the screen
Simple and intuitive user interface
Installation
To run the Clicker Game locally, follow these steps:

Clone the repository to your local machine:
bash
Copy code
git clone https://github.com/your-username/clicker-game.git
Open the project in your preferred Java IDE.
Compile and run the Gui.java file.
Usage
Click the "Click" button to increase the counter.
The number of clicks will be displayed on the screen.
Contributing
Contributions are welcome! If you find any bugs or have suggestions for improvements, feel free to open an issue or submit a pull request.